refactor(currency): consolidate currency modules into single file
- Merge currency-codes.ts into currency.ts as single source of truth
- Use correct ISO 4217 decimal digits in formatCurrency() instead of
  hardcoded 2 (fixes formatting for JPY, KRW, BHD, KWD, etc.)
- Replace inline toLocaleString() calls in advanced-asset-columns and
  inventory-value-chart with centralized formatCurrency()
- Combine both test files into currency.test.ts
- Update all imports and Prisma schema reference
